Alteração do campo peso de INTEGAR para FLOAT

+++ src/app/models/Student.js
@@ -7,7 +7,7 @@ class Student extends Model {
         name: Sequelize.STRING,
         email: Sequelize.STRING,
         idade: Sequelize.INTEGER,
-        peso: Sequelize.INTEGER,
+        peso: Sequelize.FLOAT,
         altura: Sequelize.FLOAT,
       },
       {
